add cart mechanism

+++ database/migrations/2024_06_10_093903_create_add_product_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('add_product', function (Blueprint $table) {
             $table->id();
             $table->string('product_name');
-            $table->decimal('price', 8, 2);
+            $table->bigInteger('price');
             $table->text('product_description');
             $table->string('product_image');
             $table->timestamps();
